+++ pkg/nvcdi/spec/builder.go
@@ -19,6 +19,7 @@ package spec
 import (
 	"fmt"
 
+	"github.com/NVIDIA/nvidia-container-toolkit/pkg/nvcdi/transform"
 	"github.com/container-orchestrated-devices/container-device-interface/pkg/cdi"
 	"github.com/container-orchestrated-devices/container-device-interface/specs-go"
 )
@@ -31,6 +32,7 @@ type builder struct {
 	deviceSpecs []specs.Device
 	edits       specs.ContainerEdits
 	format      string
+	noSimplify  bool
 }
 
 // newBuilder creates a new spec builder with the supplied options
@@ -74,6 +76,13 @@ func (o *builder) Build() (*spec, error) {
 			return nil, fmt.Errorf("failed to get minumum required CDI spec version: %v", err)
 		}
 		raw.Version = minVersion
+	}
+
+	if !o.noSimplify {
+		err := transform.NewSimplifier().Transform(raw)
+		if err != nil {
+			return nil, fmt.Errorf("failed to simplify spec: %v", err)
+		}
 	}
 
 	s := spec{
@@ -126,5 +135,12 @@ func WithClass(class string) Option {
 func WithFormat(format string) Option {
 	return func(o *builder) {
 		o.format = format
+	}
+}
+
+// WithNoSimplify sets whether the spec must be simplified
+func WithNoSimplify(noSimplify bool) Option {
+	return func(o *builder) {
+		o.noSimplify = noSimplify
 	}
 }