+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.13.0-rc.3
 
 * Prefer /run over /var/run when locating nvidia-persistenced and nvidia-fabricmanager sockets.
+* Only initialize NVML for modes that require it when runing `nvidia-ctk cdi generate`
 
 ## v1.13.0-rc.2